backend/pyproject.toml
[project]
name = "newsai-backend"
version = "0.1.0"
description = "NewsAI Studio - Automated News Video Platform Backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "python-dotenv>=1.0.0",
    "pyjwt>=2.8.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.9",
    "httpx>=0.27.0",
    "feedparser>=6.0.0",
    "aiohttp>=3.9.0",
    "python-jose[cryptography]>=3.3.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "aiosqlite>=0.20.0",
    "moviepy>=2.0.0",
    "pillow>=9.2.0,<12.0",
    "plotly>=5.0.0",
    "kaleido>=0.2.0",
    "google-auth>=2.0.0",
    "google-auth-oauthlib>=1.0.0",
    "google-api-python-client>=2.100.0",
    "pydantic-settings>=2.0.0",
    "edge-tts>=6.1.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
